requirements.txt
discord==1.0.1
py-rolldice==0.3.9
DnD4py==1.0.5
prometheus_client==0.8.0